refactor: separation of relp and cli from main
